Make sure namespaces in all the classes do match the corresponding project names.
Rename the class WarehouseAllOfNextHops --> WarehouseNextHops.

Delete the three classes TruckAllOf, TransferwarehouseAllOf, WarehouseAllOf, as those are not needed.

Move the implementing classes ParcelApiController and WarehouseApiController from the package services to the package services.impl.

Refactor the content of the org.openapitools package and move it to the at.fhtw.swen3 package.